requirements.txt
discord.py 
validators
